refactor: room 패키지 내 vo 패키지명을 domain 으로 변경
