+++ backend/src/main/java/com/codesquad/airbnb/room/repository/RoomRepositoryCustomImpl.java
@@ -9,11 +9,11 @@ import com.codesquad.airbnb.common.embeddable.GuestGroup;
 import com.codesquad.airbnb.common.embeddable.Location;
 import com.codesquad.airbnb.common.embeddable.StayDate;
 import com.codesquad.airbnb.reservation.Reservation.ReservationState;
+import com.codesquad.airbnb.room.domain.LocationCluster;
 import com.codesquad.airbnb.room.dto.request.RoomSearCondition;
 import com.codesquad.airbnb.room.dto.request.RoomSearCondition.PriceRange;
 import com.codesquad.airbnb.room.dto.request.RoomSearCondition.Radius;
 import com.codesquad.airbnb.room.entity.Room;
-import com.codesquad.airbnb.room.vo.LocationCluster;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
